+++ packages/frameworks/react/src/date-picker/use-date-picker.ts
@@ -6,7 +6,10 @@ import type { Optional } from '../types'
 import { useEvent } from '../use-event'
 
 export interface UseDatePickerProps
-  extends Optional<Omit<datePicker.Context, 'value' | 'focusedValue' | 'open.controlled'>, 'id'> {
+  extends Optional<
+    Omit<datePicker.Context, 'value' | 'min' | 'max' | 'focusedValue' | 'open.controlled'>,
+    'id'
+  > {
   /**
    * The initial value of the date picker
    */
@@ -19,6 +22,15 @@ export interface UseDatePickerProps
    * The value of the date picker
    */
   value?: string[]
+  /**
+   * The minimum date for the date picker in the format yyyy-mm-dd
+   */
+  min?: string
+
+  /**
+   * The maximum date for the date picker in the format yyyy-mm-dd
+   */
+  max?: string
 }
 
 export interface UseDatePickerReturn extends datePicker.Api<PropTypes> {}
@@ -30,11 +42,15 @@ export const useDatePicker = (props: UseDatePickerProps = {}): UseDatePickerRetu
     ...props,
     focusedValue: props.focusedValue ? datePicker.parse(props.focusedValue) : undefined,
     value: props.defaultValue ? datePicker.parse(props.defaultValue) : undefined,
+    max: props.max ? datePicker.parse(props.max) : undefined,
+    min: props.min ? datePicker.parse(props.min) : undefined,
     'open.controlled': props.open !== undefined,
   }
 
   const context: datePicker.Context = {
     ...initialContext,
+    max: props.max ? datePicker.parse(props.max) : undefined,
+    min: props.min ? datePicker.parse(props.min) : undefined,
     value: props.value ? datePicker.parse(props.value) : undefined,
     onValueChange: useEvent(props.onValueChange, { sync: true }),
     onFocusChange: useEvent(props.onFocusChange),