+++ packages/frameworks/solid/src/date-picker/use-date-picker.ts
@@ -5,7 +5,10 @@ import { useEnvironmentContext } from '../environment'
 import { type Optional } from '../types'
 
 export interface UseDatePickerProps
-  extends Optional<Omit<datePicker.Context, 'value' | 'focusedValue' | 'open.controlled'>, 'id'> {
+  extends Optional<
+    Omit<datePicker.Context, 'value' | 'min' | 'max' | 'focusedValue' | 'open.controlled'>,
+    'id'
+  > {
   /**
    * The focused date.
    */
@@ -14,21 +17,33 @@ export interface UseDatePickerProps
    * The value of the date picker
    */
   value?: string[]
+  /**
+   * The minimum date for the date picker in the format yyyy-mm-dd
+   */
+  min?: string
+  /**
+   * The maximum date for the date picker in the format yyyy-mm-dd
+   */
+  max?: string
 }
 export interface UseDatePickerReturn extends Accessor<datePicker.Api<PropTypes>> {}
 
 export const useDatePicker = (props: UseDatePickerProps): UseDatePickerReturn => {
-  const [local, rest] = splitProps(props, ['value', 'focusedValue'])
   const getRootNode = useEnvironmentContext()
+  const [localProps, restProps] = splitProps(props, ['value', 'focusedValue', 'min', 'max'])
+
   const context = mergeProps(
     () => ({
       id: createUniqueId(),
       getRootNode,
-      focusedValue: local.focusedValue ? datePicker.parse(local.focusedValue) : undefined,
-      value: local.value ? datePicker.parse(local.value) : undefined,
+      focusedValue: localProps.focusedValue ? datePicker.parse(localProps.focusedValue) : undefined,
+      value: localProps.value ? datePicker.parse(localProps.value) : undefined,
+      max: localProps.max ? datePicker.parse(localProps.max) : undefined,
+      min: localProps.min ? datePicker.parse(localProps.min) : undefined,
     }),
-    rest,
+    restProps,
   )
+
   const [state, send] = useMachine(datePicker.machine(context), { context })
 
   return createMemo(() => datePicker.connect(state, send, normalizeProps))